Chore: Removed JDBC entirely and cleaned up some code

+++ interfaces/src/main/java/ar/edu/itba/paw/services/DriverService.java
@@ -79,7 +79,7 @@ public interface DriverService {
 
     List<Vehicle> getVehicles(Driver driver, int page);
 
-    int getVehicleCount(Driver driver);
+    long getVehicleCount(Driver driver);
 
     void deleteVehicle(Vehicle vehicle);
 }
